rename: 패키지 구조 domain, global로 나눠서 수정
