+++ src/main/java/codesquad/cafe/domain/article/controller/ArticleController.java
@@ -1,8 +1,8 @@
-package codesquad.cafe.article.controller;
+package codesquad.cafe.domain.article.controller;
 
-import codesquad.cafe.article.dtio.ArticleRequestDto;
-import codesquad.cafe.article.dtio.ArticleResponseDto;
-import codesquad.cafe.article.service.ArticleService;
+import codesquad.cafe.domain.article.dtio.ArticleRequestDto;
+import codesquad.cafe.domain.article.dtio.ArticleResponseDto;
+import codesquad.cafe.domain.article.service.ArticleService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -25,10 +25,10 @@ public class ArticleController {
         return "index";
     }
 
-    @GetMapping("/questions/form")
-    public String showQnaForm() {
-        return "qna/form";
-    }
+//    @GetMapping("/questions/form")
+//    public String showQnaForm() {
+//        return "qna/form";
+//    }
 
     @PostMapping("/questions")
     public String writePost(@ModelAttribute ArticleRequestDto articleRequestDto) {
